requirements1.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
chardet==5.2.0
Django==5.1.7
django-cors-headers==4.3.1
django-jalali==7.0.0
django-jalali-date==1.1.3
gunicorn==21.2.0
jalali_core==1.0.0
jdatetime==5.2.0
packaging==24.2
pillow==11.1.0
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.3.1
setuptools==78.1.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
